+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Preview release
 - Removed auto-scroll as it did not work as expected on a smartphone.
 - Renamed Button "New list" to "Clear list".
 - Added and changed some texts.
+- Made the "Share" button a selecteable option.
 
 ## 0.9.0 (2022-04-14)